Add codeium once again
I have actually missed having codeium around, and the main reason for
removing it was the annoying way that the tab completion worked, and I
didn't manage to fix it. Now it seems that I got something working
though, albeit not all the way where I want it. Tab is no longer
accepting suggestions, but I can't seem to cycle between suggestions
with the keybindings that I chose. Let's see if I can fix this at a
later time.