Neovim config restructure
Having everything in a single file started to become messy, so I started
to revamp things. This makes it more modular, and it's easier to keep
track of what goes where. Previously, the config for a particular
plugin could be anywhere, but now it's going to be much more easy to
find and maintain.